+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## v4.13.0 (21/02/2022)
+- [**closed**] #13346 商品種類区分による検索を追加 [#1769](https://mygithub.ritscm.xyz/cotos/CommonLibs/pull/1769)
+- [**closed**] メールアドレス情報.テーブル区分を追加 [#1863](https://mygithub.ritscm.xyz/cotos/CommonLibs/pull/1863)
+- [**closed**] #13376_更新案内メール送信フラグを販売店に追加 [#1854](https://mygithub.ritscm.xyz/cotos/CommonLibs/pull/1854)
+- [**closed**] #13882 バッチ用メール送信エラーメッセージ追加 [#1849](https://mygithub.ritscm.xyz/cotos/CommonLibs/pull/1849)
+- [**closed**] #13505 品種マスタ.拡張項目を追加 [#1848](https://mygithub.ritscm.xyz/cotos/CommonLibs/pull/1848)
+
+---
+
+## v4.12.0 (07/02/2022)
+- [**closed**] #12964 外部API呼び出し用Interceptor修正 [#1845](https://mygithub.ritscm.xyz/cotos/CommonLibs/pull/1845)
+- [**closed**] #12964 TrendMicro連携処理修正 [#1837](https://mygithub.ritscm.xyz/cotos/CommonLibs/pull/1837)
+- [**closed**] #12964 TrendMicro連携処理修正 [#1833](https://mygithub.ritscm.xyz/cotos/CommonLibs/pull/1833)
+- [**closed**] #12964 外部API呼び出し用RestTemplate作成 [#1826](https://mygithub.ritscm.xyz/cotos/CommonLibs/pull/1826)
+
+---
+
 ## v4.11.0 (24/01/2022)
 - [**closed**] #13460 Umbrella対応のライセンス関連DTOの修正 [#1784](https://mygithub.ritscm.xyz/cotos/CommonLibs/pull/1784)
 - [**closed**] #13510_JsonUtilにアノテーションを追加 [#1791](https://mygithub.ritscm.xyz/cotos/CommonLibs/pull/1791)